requirements.txt
annotated-types==0.7.0
anyio==4.4.0
exceptiongroup==1.2.2
fastapi==0.112.0
idna==3.7
pillow==10.4.0
pip==21.2.3
pydantic==2.8.2
pydantic-core==2.20.1
pymupdf==1.24.9
pymupdfb==1.24.9
setuptools==57.4.0
sniffio==1.3.1
starlette==0.37.2
typing-extensions==4.12.2
